Added expandable table cells and refactored table to use HTML table instead of flexbox.
