+++ components/Table/actions/index.ts
@@ -1,5 +1,3 @@
-export { default as FilterAction } from "./FilterAction";
-export { default as SortAction } from "./SortAction";
-export { default as PaginationAction } from "./PaginationAction";
-export { default as RefreshAction } from "./RefreshAction";
-export { default as DeleteAction } from "./DeleteAction";
+export * from "./GlobalActions";
+export * from "./TableActions";
+export * from "./SortColumn";
