+++ components/Table/hooks/useTable.ts
@@ -1,4 +1,5 @@
 import {
+  FilterFn,
   getCoreRowModel,
   getExpandedRowModel,
   getFilteredRowModel,
@@ -9,17 +10,16 @@ import {
   TableOptions,
   useReactTable,
 } from "@tanstack/react-table";
-import { ColumnTypeMeta } from "../types";
-
-type ColumnFormatterMeta = {
-  [key: string]: (value: unknown) => string;
-};
+import { ColumnTypeMeta, ColumnFormatterMeta } from "../types";
+import { rankItem } from "@tanstack/match-sorter-utils";
 
 declare module "@tanstack/react-table" {
   // eslint-disable-next-line
   interface TableMeta {
     formatter: ColumnFormatterMeta;
     columnType: ColumnTypeMeta;
+    rowType: "data" | "expand";
+    getExpandRows: (row: any) => any[];
   }
 }
 
@@ -31,6 +31,7 @@ type UseTableOptions<TData extends RowData> = Omit<
   | "getFilteredRowModel"
   | "getGroupedRowModel"
   | "getSortedRowModel"
+  | "getSubRows"
 > & {
   usePagination?: boolean;
   useExpanded?: boolean;
@@ -38,11 +39,22 @@ type UseTableOptions<TData extends RowData> = Omit<
   useSorted?: boolean;
   // formatter can be used to format a cell's value.
   // Its key must be the column id for that cell
-  formatter?: ColumnFormatterMeta;
+  formatter: ColumnFormatterMeta;
   columnType: ColumnTypeMeta;
+  getSubRows?: (row: any) => any[];
 };
 
 type UseTableReturn<TData extends RowData> = Table<TData>;
+
+const globalFilterFn: FilterFn<any> = (row, columnId, filterValue, addMeta) => {
+  const itemRank = rankItem(row.getValue(columnId), filterValue);
+
+  addMeta({
+    itemRank,
+  });
+
+  return itemRank.passed;
+};
 
 export function useTable<TData extends RowData>(
   options: UseTableOptions<TData>
@@ -54,6 +66,7 @@ export function useTable<TData extends RowData>(
     useFilter = true,
     formatter,
     columnType,
+    getSubRows,
     ...tableOptions
   } = options;
 
@@ -61,9 +74,13 @@ export function useTable<TData extends RowData>(
     ...tableOptions,
     meta: {
       ...tableOptions.meta,
+      rowType: useExpanded ? "expand" : "data",
+      getExpandRows: getSubRows ?? (() => []),
       formatter: formatter ?? {},
       columnType,
     },
+    manualExpanding: useExpanded ? true : undefined,
+    globalFilterFn: globalFilterFn,
     getCoreRowModel: getCoreRowModel(),
     getSortedRowModel: useSorted ? getSortedRowModel() : undefined,
     getExpandedRowModel: useExpanded ? getExpandedRowModel() : undefined,
