+++ components/Table2/SimpleTable.tsx
