+++ components/Table2/actions/index.ts
@@ -1,3 +1,5 @@
-export * from "./GlobalActions";
-export * from "./TableActions";
-export * from "./SortColumn";
+export { default as FilterAction } from "./FilterAction";
+export { default as SortAction } from "./SortAction";
+export { default as PaginationAction } from "./PaginationAction";
+export { default as RefreshAction } from "./RefreshAction";
+export { default as DeleteAction } from "./DeleteAction";
