+++ package.json
@@ -20,6 +20,7 @@
     "@mui/lab": "^5.0.0-alpha.89",
     "@mui/material": "^5.8.1",
     "@mui/styled-engine-sc": "^5.8.0",
+    "@tanstack/match-sorter-utils": "^8.1.1",
     "@tanstack/react-table": "^8.3.3",
     "@visx/axis": "^2.11.1",
     "@visx/event": "^2.6.0",
