+++ pages/index.tsx
@@ -10,14 +10,13 @@ import { ControlledSelect, Input, Select } from "components/base";
 import { FormProvider, useForm } from "react-hook-form";
 import { DatePicker, TimePicker } from "components/base/Pickers";
 import { useDateTime } from "common/hooks";
-import { Table } from "components/Table2";
-import { useColumnDef, useTable } from "components/Table2/hooks";
+import { useColumnDef, useTable, Table } from "components/Table";
 import { useQuery } from "react-query";
 import { fetch, getAllLocations, QueryKeys } from "api";
 import { useEffect, useMemo } from "react";
 
 type LocationEntity = {
-  Name: string;
+  name: string;
   uid: number;
 };
 
@@ -30,7 +29,7 @@ const Home: NextPage = () => {
         if (data) {
           return data.map((d) => ({
             uid: d.uid,
-            Name: d.location_name,
+            name: d.location_name,
           }));
         }
       },
@@ -42,7 +41,7 @@ const Home: NextPage = () => {
       return allLocations.reduce((acc, curr) => {
         acc[curr.uid] = curr;
         return acc;
-      }, {} as { [key: number]: { Name: string; uid: number } });
+      }, {} as { [key: number]: { name: string; uid: number } });
     }
   }, [allLocations]);
 
@@ -60,9 +59,10 @@ const Home: NextPage = () => {
       {
         id: "name",
         type: "input",
-        name: "Name",
+        header: "Name",
+        inputName: "name",
         placeholder: "Enter a location",
-        accessorKey: "Name",
+        accessorKey: "name",
       },
     ],
   });
